Restructure helper_functions into visualisation functions
