+++ tests/visualization_test.py
@@ -1,0 +1,2 @@
+#sanity tests for the visualization functions
+# TODO: Define tests for show_my_power_spectrum
